Moved classes wrongly placed in Receiver package
