+++ QKSMS/src/main/java/com/moez/QKSMS/receiver/MessagingReceiver.java
@@ -16,6 +16,7 @@ import com.moez.QKSMS.data.Message;
 import com.moez.QKSMS.service.NotificationService;
 import com.moez.QKSMS.common.NotificationManager;
 import com.moez.QKSMS.common.SmsHelper;
+import com.moez.QKSMS.service.UnreadBadgeService;
 import com.moez.QKSMS.ui.settings.SettingsFragment;
 import org.mistergroup.muzutozvednout.ShouldIAnswerBinder;
