--- QKSMS/src/main/java/com/moez/QKSMS/receiver/QKIntent.java
@@ -1,5 +1,0 @@
-package com.moez.QKSMS.receiver;
-
-public class QKIntent {
-    public static final String ACTION_ANALYTICS_REPORT = "com.moez.QKSMS.action.analytics_report";
-}
